+++ tests/testapp/settings.py
@@ -1,5 +1,6 @@
 import os
 
+
 DATABASES = {"default": {"ENGINE": "django.db.backends.sqlite3", "NAME": ":memory:"}}
 
 INSTALLED_APPS = [